Move interfaces/reducers outside of components to allow simpler component grouping
