+++ client/src/interfaces/IGroup.ts
