+++ client/src/interfaces/ITaskState.ts
